requirements.txt
# Core PyTorch (MPS support for macOS)
# PyTorch 2.0+ has native MPS support for Apple Silicon
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision and Detection
ultralytics>=8.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0

# Graph Neural Networks and Transformers
torch-geometric>=2.3.0
transformers>=4.30.0
timm>=0.9.0

# CoreML for model conversion and deployment
coremltools>=7.0.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Experiment Tracking (Optional)
wandb>=0.15.0
tensorboard>=2.13.0

# Additional utilities
tqdm>=4.65.0
pillow>=9.0.0
pyyaml>=6.0

# Web/HTTP (for dataset download)
requests>=2.31.0
